Move UI navigation tests to separate CI job with ui_test tag
Extract integration tests from ui_test.go into ui_integration_test.go
behind a //go:build ui_test tag, matching the pattern used by docker,
idp, and openbao integration suites. Add a parallel ui CI job with
its own coverage output wired into the coverage merge step.
