+++ .github/workflows/ci.yml
@@ -17,6 +17,7 @@ on:
           - docker
           - idp
           - openbao
+          - ui
 
 concurrency:
   group: ci-${{ github.event.pull_request.number || github.ref }}
@@ -119,10 +120,25 @@ jobs:
           name: coverage-openbao
           path: coverage-openbao.out
 
+  ui:
+    runs-on: ubuntu-latest
+    needs: [check]
+    if: github.event_name != 'merge_group' && github.event_name != 'push' && (inputs.job == '' || inputs.job == 'ui')
+    steps:
+      - uses: actions/checkout@v6
+      - uses: actions/setup-go@v6
+        with:
+          go-version-file: go.mod
+      - run: go test -tags ui_test -coverprofile=coverage-ui.out -coverpkg=./internal/... ./internal/ui/...
+      - uses: actions/upload-artifact@v7
+        with:
+          name: coverage-ui
+          path: coverage-ui.out
+
   coverage:
     runs-on: ubuntu-latest
     if: github.event_name != 'merge_group' && github.event_name != 'push' && inputs.job == ''
-    needs: [check, docker, idp, openbao]
+    needs: [check, docker, idp, openbao, ui]
     steps:
       - uses: actions/checkout@v6
       - uses: actions/setup-go@v6
@@ -145,6 +161,6 @@ jobs:
       - uses: codecov/codecov-action@v5
         if: always()
         with:
-          files: coverage-check.out,coverage-docker.out,coverage-idp.out,coverage-openbao.out
-          flags: unit,docker,idp,openbao
+          files: coverage-check.out,coverage-docker.out,coverage-idp.out,coverage-openbao.out,coverage-ui.out
+          flags: unit,docker,idp,openbao,ui
           fail_ci_if_error: false
